move git related code and package to cli, update tests so that core and cli tests only depend on itself and/or core and idmtools-test packages
