+++ idmtools_cli/tests/test_git_tags.py
@@ -1,12 +1,15 @@
 import os
-import sys
+from functools import partial
+
 import allure
 import pytest
+from idmtools_test.utils.test_task import TestTask
+
 from idmtools.builders import SimulationBuilder
 from idmtools.core.platform_factory import Platform
 from idmtools.entities.experiment import Experiment
-from idmtools_models.python.json_python_task import JSONConfiguredPythonTask
 from idmtools_test.utils.itest_with_persistence import ITestWithPersistence
+
 
 model_path = os.path.abspath(os.path.join("..", "..", "examples", "python_model", "inputs", "simple_python", "model.py"))
 
@@ -30,23 +33,29 @@ class TestGitTags(ITestWithPersistence):
         os.environ['IDMTOOLS_GIT_TAG_ADD_TO_EXPERIMENTS'] = 'n'
 
     def test_adding_git_tag(self):
-        base_task = JSONConfiguredPythonTask(script_path=model_path, python_path=sys.executable)
+        task = TestTask()
+        platform = Platform('TestExecute', missing_ok=True)
         builder = SimulationBuilder()
-        builder.add_sweep_definition(JSONConfiguredPythonTask.set_parameter_partial("Run_Number"), [i for i in range(5)])
-        exp = Experiment.from_builder(builder, base_task=base_task)
-        exp.run(wait_until_done=True, add_git_tags_to_simulation=True)
+        def param_update(simulation, param, value):
+            return simulation.task.set_parameter(param, value)
 
-        self.assertTrue(exp.succeeded)
+        builder.add_sweep_definition(partial(param_update, param="Run_Number"), [i for i in range(5)])
+        exp = Experiment.from_builder(builder, base_task=task)
+        exp.run(platform=platform, wait_until_done=True, add_git_tags_to_simulation=True)
+
+        self.assertFalse(exp.succeeded)
         self.assertIn('git_hash', exp.tags)
         self.assertNotIn('git_hash', exp.simulations[0].tags)
 
     def test_adding_git_tag_as_option_to_run(self):
-        base_task = JSONConfiguredPythonTask(script_path=model_path, python_path=sys.executable)
+        task = TestTask()
         builder = SimulationBuilder()
-        builder.add_sweep_definition(JSONConfiguredPythonTask.set_parameter_partial("Run_Number"), [i for i in range(5)])
-        exp = Experiment.from_builder(builder, base_task=base_task)
+        def param_update(simulation, param, value):
+            return simulation.task.set_parameter(param, value)
+        builder.add_sweep_definition(partial(param_update, param="Run_Number"), [i for i in range(5)])
+        exp = Experiment.from_builder(builder, base_task=task)
         exp.run(wait_until_done=True, add_git_tags_to_simulations=True)
 
-        self.assertTrue(exp.succeeded)
+        self.assertFalse(exp.succeeded)
         self.assertIn('git_hash', exp.tags)
         self.assertIn('git_hash', exp.simulations[0].tags)
