+++ idmtools_cli/tests/test_github_url_parse.py
@@ -4,6 +4,7 @@ import unittest
 from unittest import TestCase
 
 import pytest
+# from idmtools_cli.utils.gitrepo import GitRepo
 from idmtools.utils.gitrepo import GitRepo
