+++ idmtools_core/build_requirements.txt
@@ -5,3 +5,4 @@ flake8-docstrings>=1.6.0
 idm-buildtools>=1.0.3
 twine>=3.4.1
 jinja2~=3.1.2
+pygit2>=1.4.0
