+++ idmtools_core/requirements.txt
@@ -12,4 +12,4 @@ tabulate>=0.8.9,<0.10
 tqdm>=4.52.0,<5
 jinja2~=3.1.2
 packaging>=20.4,<22.0
-pygit2>=1.4.0
+
