+++ idmtools_core/tests/test_platform_factory.py
@@ -1,6 +1,5 @@
 import allure
 import os
-import unittest.mock
 import pytest
 from dataclasses import fields
 from idmtools.config import IdmConfigParser
@@ -57,56 +56,11 @@ class TestPlatformFactory(ITestWithPersistence):
         except Exception as ex:
             self.assertIn("__new__() missing 1 required positional argument: 'block'", ex.args[0])
 
-    @pytest.mark.comps
-    @pytest.mark.timeout(60)
-    @unittest.mock.patch('idmtools_platform_comps.comps_platform.COMPSPlatform._login', side_effect=lambda: True)
-    @allure.story("COMPS")
     @allure.story("Configuration")
-    def test_create_from_block(self, mock_login):
-        p2 = Platform('COMPS')
-        self.assertEqual(p2.__class__.__name__, 'COMPSPlatform')
-        self.assertEqual(mock_login.call_count, 1)
-
-        p3 = Platform('Test')
-        self.assertEqual(p3.__class__.__name__, 'TestPlatform')
-        p3.cleanup()
-        del p2
-
-    @pytest.mark.comps
-    @pytest.mark.timeout(60)
-    @unittest.mock.patch('idmtools_platform_comps.comps_platform.COMPSPlatform._login', side_effect=lambda: True)
-    @allure.story("COMPS")
-    @allure.story("Configuration")
-    def test_platform_factory(self, mock_login):
-        platform1 = Platform('COMPS')
-        self.assertEqual(platform1.__class__.__name__, 'COMPSPlatform')
-        self.assertEqual(mock_login.call_count, 1)
-
-        platform3 = Platform('Test')
-        self.assertEqual(platform3.__class__.__name__, 'TestPlatform')
-        platform3.cleanup()
-
-    @pytest.mark.comps
-    @unittest.mock.patch('idmtools_platform_comps.comps_platform.COMPSPlatform._login', side_effect=lambda: True)
-    @allure.story("COMPS")
-    @allure.story("Configuration")
-    def test_COMPSPlatform(self, mock_login):
-        platform = Platform('COMPS')
-        self.assertEqual(mock_login.call_count, 1)
-        members = platform.__dict__
-
-        field_name = {f.name for f in fields(platform) if f.init}
-        keys = field_name.intersection(members.keys())
-        kwargs = {key: members[key] for key in keys if not key.startswith('_')}
-
-        platform2 = Platform('COMPS', **kwargs)
-        self.assertEqual(mock_login.call_count, 2)
-        # Note, due to new bug fix for platform uid(in iitem.py), we can not longer compare 2 platforms equal as before.
-        # We used to have _uid member in platform which was comparison field and always None. Now, it has value with bug
-        # fix. So in order to compare 2 platforms equal, let's ignore this property and compare rest of members
-        platform.uid = None
-        platform2.uid = None
-        self.assertEqual(platform, platform2)
+    def test_create_from_block(self):
+        p = Platform('Test')
+        self.assertEqual(p.__class__.__name__, 'TestPlatform')
+        p.cleanup()
 
     def test_TestPlatform(self):
         platform = Platform('Test')
