+++ idmtools_test/idmtools_test/inputs/configuration/idmtools_test.ini
@@ -64,3 +64,8 @@ console = on
 [Custom_Local]
 type = Local
 
+[Test]
+type = Test
+num_retries = 5
+num_cores = 1
+
